Feat [#5] 마이페이지 header 분리 및 backButton 로직 추가
